+++ src/main/java/com/global/api/entities/reporting/SurchargeLookup.java
@@ -1,0 +1,13 @@
+package com.global.api.entities.reporting;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class SurchargeLookup {
+    private String gatewayTxnId;
+    private String gatewayRspCode;
+    private String gatewayRspMsg;
+    private String isSurchargeable;
+}