+++ src/main/java/com/global/api/terminals/upa/Entities/Enums/UpaMessageId.java
@@ -24,5 +24,7 @@ public enum UpaMessageId {
     StartCardTransaction,
     GetSignature,
     DeleteSAF,
-    RegisterPOS
+    RegisterPOS,
+
+    PrintData
 }